requirements.txt
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
pyjwt>=2.1.0
cryptography>=3.4.0
open3d>=0.13.0
jupyter>=1.0.0
notebook>=6.4.0
seaborn>=0.11.0
plotly>=5.0.0
python-binance>=1.0.15
ccxt>=1.60.0 